requirements.txt
pandas == 1.2.4
kafka-python == 2.0.2
psycopg2 == 2.8.6
SQLAlchemy == 1.4.17
multiprocessing == 2.6.2.1
pendulum == 2.1.2
Flask == 1.1.2
Flask-RESTful == 0.3.8
Flask-SQLAlchemy == 2.4.3
pylint == 2.5.2
pylint-flask == 0.6
pylint_flask_sqlalchemy == 0.2.0